LayoutTests/fast/css-grid-layout/grid-shorthand-get-set.html

<!DOCTYPE html>
<html>
<head>
<link href="resources/grid.css" rel="stylesheet">
<style>
#gridWithDense {
    grid: dense;
}
#gridWithTemplate {
    grid: 10px / 15px;
}
#gridWithTemplate1 {
    grid: 10px / none;
}
#gridWithTemplateNone {
    grid: 10px / none;
}
#gridWithAutoFlowDenseAndRowsAndColumns {
    grid: dense 10px / 20px;
}
#gridWithAutoFlowAndRows {
    grid: column 10px;
}
#gridWithAutoFlowDense {
    grid: dense 10px;
}
#gridWithAutoFlowColumnDense {
    grid: column dense 10px;
}
#gridWithAutoFlowDenseRow {
    grid: dense row 10px;
}
#gridWithAutoFlowAndRowsAndColumns {
    grid: column 10px / 20px;
}

/* Bad values. */

#gridWithNone {
    grid: none;
}
#gridWithAutoFlowNoneAndRowsAndColumns {
    grid: none 10px / 20px;
}
#gridWithAutoFlowNone {
    grid: none 10px;
}
#gridWithExplicitAndImplicit {
    grid: 10px / 20px column;
}
#gridWithMisplacedNone1 {
    grid: column 10px / none 20px;
}
#gridWithMisplacedNone2 {
    grid: 10px / 20px none;
}
#gridWithWrongSlash1 {
    grid: column / 10px;
}
#gridWithWrongSlash2 {
    grid: column 10px 20px;
}
#gridWithWrongSlash3 {
    grid: 5x / 10px / 20px;
}
#gridWithAutoFlowRowAndColumn {
    grid: row column 10px;
}
#gridWithoutAutoFlowAndExtraBreath {
    grid: row 10px / 20px 30px;
}
#gridWithAutoFlowString1 {
    grid: row 10px "a";
}
#gridWithAutoFlowString2 {
    grid: row "a" 10px;
}
#gridWithAutoFlowString3 {
    grid: row "a";
}
#gridWithTemplateAndAutoFlow {
    grid: 10px / 10px row;
}
#gridWithTemplateAndMisplacedString1 {
    grid: 10px / 10px "a";
}
#gridWithMisplacedDense {
    grid: dense column dense;
}
</style>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div class="grid" id="gridWithDense"></div>
<div class="grid" id="gridWithTemplate"></div>
<div class="grid" id="gridWithTemplate1"></div>
<div class="grid" id="gridWithAutoFlowDenseAndRowsAndColumns"></div>
<div class="grid" id="gridWithAutoFlowAndRows"></div>
<div class="grid" id="gridWithAutoFlowDense"></div>
<div class="grid" id="gridWithAutoFlowColumnDense"></div>
<div class="grid" id="gridWithAutoFlowDenseRow"></div>
<div class="grid" id="gridWithAutoFlowAndRowsAndColumns"></div>
<div class="grid" id="gridWithNone"></div>
<div class="grid" id="gridWithTemplateNone"></div>
<div class="grid" id="gridWithAutoFlowNoneAndRowsAndColumns"></div>
<div class="grid" id="gridWithAutoFlowNone"></div>
<div class="grid" id="gridWithExplicitAndImplicit"></div>
<div class="grid" id="gridWithMisplacedNone1"></div>
<div class="grid" id="gridWithMisplacedNone2"></div>
<div class="grid" id="gridWithWrongSlash1"></div>
<div class="grid" id="gridWithWrongSlash2"></div>
<div class="grid" id="gridWithWrongSlash3"></div>
<div class="grid" id="gridWithAutoFlowRowAndColumn"></div>
<div class="grid" id="gridWithoutAutoFlowAndExtraBreath"></div>
<div class="grid" id="gridWithAutoFlowString1"></div>
<div class="grid" id="gridWithAutoFlowString2"></div>
<div class="grid" id="gridWithAutoFlowString3"></div>
<div class="grid" id="gridWithTemplateAndAutoFlow"></div>
<div class="grid" id="gridWithTemplateAndMisplacedString1"></div>
<div class="grid" id="gridWithMisplacedDense"></div>
<script src="resources/grid-shorthand-parsing-utils.js"></script>
<script>
    description("This test checks that the 'grid' shorthand is properly parsed and the longhand properties correctly assigned.");

    debug("Test getting the longhand values when shorthand is set through CSS.");
    testGridDefinitionsValues(document.getElementById("gridWithDense"), "none", "none", "none", "row dense", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithTemplate"), "15px", "10px", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithTemplate1"), "none", "10px", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithTemplateNone"), "none", "10px", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowDenseAndRowsAndColumns"), "none", "none", "none", "row dense", "20px", "10px");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowAndRows"), "none", "none", "none", "column", "10px", "10px");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowDense"), "none", "none", "none", "row dense", "10px", "10px");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowColumnDense"), "none", "none", "none", "column dense", "10px", "10px");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowDenseRow"), "none", "none", "none", "row dense", "10px", "10px");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowAndRowsAndColumns"), "none", "none", "none", "column", "20px", "10px");

    debug("");
    debug("Test setting wrong values for 'grid' shorthand through CSS (they should resolve to the default: 'row')");
    testGridDefinitionsValues(document.getElementById("gridWithNone"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowNoneAndRowsAndColumns"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowNone"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithExplicitAndImplicit"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithMisplacedNone1"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithMisplacedNone2"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithMisplacedDense"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithWrongSlash1"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithWrongSlash2"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithWrongSlash3"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowRowAndColumn"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithoutAutoFlowAndExtraBreath"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowString1"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowString2"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithAutoFlowString3"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithTemplateAndAutoFlow"), "none", "none", "none", "row", "auto", "auto");
    testGridDefinitionsValues(document.getElementById("gridWithTemplateAndMisplacedString1"), "none", "none", "none", "row", "auto", "auto");

    debug("");
    debug("Test getting and setting 'grid' shorthand through JS");
    testGridDefinitionsSetJSValues("20px / 10px", "10px", "20px", "none", "row", "auto", "auto", "10px", "20px", "none", "initial", "initial", "initial");
    testGridDefinitionsSetJSValues("[line] 'a' 20px / 10px", "10px", "[line] 20px", "\"a\"", "row", "auto", "auto", "10px", "[line] 20px", "\"a\"", "initial", "initial", "initial");
    testGridDefinitionsSetJSValues("row dense 20px", "none", "none", "none", "row dense", "20px", "20px", "initial", "initial", "initial", "row dense", "20px", "20px");
    testGridDefinitionsSetJSValues("column 20px / 10px", "none", "none", "none", "column", "10px", "20px", "initial", "initial", "initial", "column", "10px", "20px");

    debug("");
    debug("Test the initial value");
    var element = document.createElement("div");
    document.body.appendChild(element);
    testGridDefinitionsValues(element, "none", "none", "none", "row", "auto", "auto");
    shouldBe("getComputedStyle(element, '').getPropertyValue('grid-template-columns')", "'none'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('grid-template-rows')", "'none'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('grid-template-areas')", "'none'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('grid-auto-flow')", "'row'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('grid-auto-columns')", "'auto'");
    shouldBe("getComputedStyle(element, '').getPropertyValue('grid-auto-rows')", "'auto'");

    debug("");
    debug("Test setting grid-template-columns and grid-template-rows back to 'none' through JS");
    testGridDefinitionsSetJSValues("column 10px / 20px", "none", "none", "none", "column", "20px", "10px", "initial", "initial", "initial", "column", "20px", "10px");
    testGridDefinitionsSetJSValues("none", "none", "none", "none", "row", "auto", "auto", "none", "none", "none", "initial", "initial", "initial");

    debug("");
    debug("Test the inherit value on reset-only subproperties (grid-*-gap)");
    document.body.style.gridRowGap = "100px";
    document.body.style.gridColumnGap = "20px";
    var anotherElement = document.createElement("div");
    document.body.appendChild(anotherElement);
    shouldBeEqualToString("getComputedStyle(anotherElement, '').getPropertyValue('grid-column-gap')", "0px");
    shouldBeEqualToString("getComputedStyle(anotherElement, '').getPropertyValue('grid-row-gap')", "0px");
    anotherElement.style.grid = "inherit";
    shouldBeEqualToString("getComputedStyle(anotherElement, '').getPropertyValue('grid-column-gap')", "20px");
    shouldBeEqualToString("getComputedStyle(anotherElement, '').getPropertyValue('grid-row-gap')", "100px");
</script>
</body>
</html>
